requirements.txt
tensorflow
fastapi
uvicorn
python-multipart
pillow
tensorflow-serving-api==2.5.0
matplotlib
numpy
streamlit 
requests
python-dotenv
gpt_index==0.4.15
langchain
google-generativeai
langchain-google-genai 
streamlit_chat
python-box==6.0.2
joblib
dvc
python-dotenv
langchain_community